@@ -1,18 +1,27 @@
 {
   "name": "phosphor-react",
-  "version": "1.4.1",
+  "version": "2.0.0",
   "description": "A clean and friendly icon family for React",
-  "author": "rektdeckard",
+  "author": {
+    "name": "Tobias Fried",
+    "email": "friedtm@gmail.com",
+    "url": "https://tobiasfried.com"
+  },
   "license": "MIT",
-  "repository": "phosphor-icons/phosphor-react",
+  "repository": "phosphor-icons/react",
   "homepage": "https://phosphoricons.com",
-  "module": "dist/index.esm.js",
-  "main": "dist/index.js",
   "sideEffects": false,
-  "typings": "dist/index.d.ts",
+  "module": "./dist/index.es.js",
+  "main": "./dist/index.umd.js",
+  "exports": {
+    ".": {
+      "import": "./dist/index.es.js",
+      "require": "./dist/index.umd.js"
+    }
+  },
+  "types": "./dist/index.d.ts",
   "files": [
-    "dist",
-    "src"
+    "./dist"
   ],
   "engines": {
     "node": ">=10"
@@ -28,40 +37,29 @@
     "UX"
   ],
   "scripts": {
-    "start": "tsdx watch",
-    "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "lint": "tsdx lint",
-    "prepare": "tsdx build",
+    "serve": "vite serve ./example",
+    "build": "vite build && tsc --emitDeclarationOnly",
+    "test": "vitest",
     "collate": "node bin/collate.js",
     "assemble": "node bin/assemble.js",
-    "format": "prettier --write \"./src/**/*.{js,jsx,ts,tsx,json,vue}\""
+    "format": "prettier --write \"./src/**/*.{js,jsx,ts,tsx,json,vue}\" \"./example/**/*.{js,jsx,ts,tsx,json,vue}\" \"./bin/**/*.{js,jsx,ts,tsx,json,vue}\""
   },
   "peerDependencies": {
     "react": ">=16"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "tsdx lint"
-    }
-  },
-  "prettier": {
-    "printWidth": 80,
-    "semi": true,
-    "singleQuote": false,
-    "trailingComma": "es5"
-  },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^10.4.9",
-    "@types/react": "^16.9.43",
-    "@types/react-dom": "^16.9.8",
+    "@testing-library/react": "^13.4.0",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react": "^3.0.1",
     "chalk": "^4.1.0",
-    "husky": "^4.2.5",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "tsdx": "^0.14.1",
-    "tslib": "^2.0.0",
-    "typescript": "^3.9.7"
+    "jsdom": "^21.1.0",
+    "prettier": "^2.8.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4",
+    "vitest": "^0.28.2"
   }
 }
